+++ src/readData.h
@@ -30,7 +30,7 @@ Header Guards
 
 struct dataStruct
 {
-	int id;  // Line Number
+	long long int id;  // Line Number
 	double coordinates[3];     // column 1
 };
